Player picks up when a song is added

+++ lib/playlist.rb
@@ -13,6 +13,7 @@ module SonicOMatic
     def next_random
       reload
       track_info = @queue.sample
+      return nil unless track_info
       track_code, tweet, user = track_info.split(SonicOMatic::DIVISOR)
       print_fancy_track_summary(tweet, user)
       track_code
@@ -21,17 +22,20 @@ module SonicOMatic
     def next(index = 0)
       reload
       track_info = @queue.count > index ? @queue[index] : nil
+      return nil unless track_info
       track_code, tweet, user = track_info.split(SonicOMatic::DIVISOR)
       print_fancy_track_summary(tweet, user)
       track_code
     end
 
     def reload
-      playlist_file = File.open(@file_name, 'r')
       @queue = []
-      playlist_file.each {|line| @queue << line }
-      playlist_file.close
-      puts "  #{SonicOMatic::DIVISOR} #{queue.count} tracks in #{@file_name}"
+      if File.exists?(@file_name)
+        playlist_file = File.open(@file_name, 'r')
+        playlist_file.each {|line| @queue << line }
+        playlist_file.close
+        puts "  #{SonicOMatic::DIVISOR} #{queue.count} tracks in #{@file_name}"
+      end
     end
 
     def print_fancy_track_summary(tweet, user)
